Initial MAVlink integration
- Decided to try out MAVlink so in this commit I have integrated a first
part of MAVlink protocol on ASAC. Initially I thought of creating a
custom dialect, but think I will go for a common one instead, we'll
see.
- Added mavlink C headers
- Created python serial client that uses pymavlink to parse the data.
The reason for not using pymavlink is that it's specialized for
ardupilot, and did not work well with rp2040.
- Added script that generates the mavlink files from xml.
This script uses absolute paths though, which should be changed in
the future