Created option to run firmware on linux
In order to aid development, it would be nice to run
certain pieces of code on a normal linux machine. For instance
if we want to run unit tests or simulations. For this, I created a
define, "LINUX" which (if defined) should make the firmware
possible to run on linux instead of rp2040.