Pid: Cleanup and refactor
I feel like this implementation is more clean and simpler.
We skip everything with timing-related I-term antiwindup and simply
let the pid control take a bool if it should ignore I term or not
This way it's easy for the callers to tell it to prevent I-term buildup (we still hav `imax` the limits the value of I-term.

Also, the pid should now not be tied to the flight control  (which it was before, due to throttle argument)